LPS-116679 Move CSV writers from DataFactory to a new class and make it available in the Freemarker context
